refactor: survey/create -> survey/general/create
