+++ src/components/surveys/create/general/SurveyCreateItem.jsx
@@ -4,11 +4,11 @@ import {
   surveyListState,
   detailMCQuestionState,
   createdQuestionCount,
-} from "../../../atoms";
+} from "../../../../atoms";
 import DetailMCQuestionCreator from "./DetailMCQuestionCreator";
 import DetailSurveyItem from "./DetailSurveyItem";
 import DeleteIcon from "@mui/icons-material/Delete";
-import { Question, MInput, Num } from "../common/styled";
+import { Question, MInput, Num } from "../../common/styled";
 
 export default function SurveyItem({ item }) {
   const [questionCount, setQuestionCount] =
