+++ src/pages/surveys/create/index.js
@@ -1,5 +1,5 @@
 import { React, useEffect } from "react";
-import CreateSurvey from "../../../components/surveys/create/CreateSurvey";
+import CreateSurvey from "../../../components/surveys/create/general/CreateSurvey";
 import { useRecoilValue } from "recoil";
 import { userState } from "../../../authentication/userState";
